// ImposterGame/ImposterGame.App/Imposter/src/server/model/iRoundParticipant.ts
/**
 * Imposter API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAccusation } from './iAccusation';
import { IPlayer } from './iPlayer';


export interface IRoundParticipant { 
    player?: IPlayer;
    readonly isImposter?: boolean;
    answer?: string;
    readonly hasAnswered?: boolean;
    accusation?: IAccusation;
    scoredPoints?: number;
}